refactoring, clearing
