--- product-detail-page/src/main/java/com/rewedigital/examples/msintegration/productdetailpage/infrastructure/eventing/ConsumerTopicConfig.java
@@ -1,10 +1,0 @@
-package com.rewedigital.examples.msintegration.productdetailpage.infrastructure.eventing;
-
-
-public interface ConsumerTopicConfig {
-
-    String getName();
-
-    boolean isPayloadSensitive();
-
-}
