+++ product-detail-page/src/main/java/com/rewedigital/examples/msintegration/productdetailpage/infrastructure/eventing/configuration/ConsumerTopicConfig.java
@@ -1,0 +1,10 @@
+package com.rewedigital.examples.msintegration.productdetailpage.infrastructure.eventing.configuration;
+
+
+public interface ConsumerTopicConfig {
+
+    String getName();
+
+    boolean isPayloadSensitive();
+
+}
