--- product-detail-page/src/main/java/com/rewedigital/examples/msintegration/productdetailpage/product/ProductEventType.java
@@ -1,15 +1,0 @@
-package com.rewedigital.examples.msintegration.productdetailpage.product;
-
-public enum ProductEventType {
-    PRODUCT_CREATED("product-created"), PRODUCT_UPDATED("product-updated");
-
-    private String name;
-
-    ProductEventType(final String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
